fix(svelte-scoped): move tests out of source files into separate test files
